W177b: near-miss batch A closes three, and the $at prediction is confirmed exactly.
  btlTarget 002D8390  628/640 nd2 -> MATCH
  fcl_misc  003C8B50  552/560 nd2 -> MATCH
  bp_tex    0021C9F0  556/560 nd5 -> MATCH

**002D8390 confirms the $at rule verbatim.** At 0x34/0x38 we emitted `slti $v0,$s0,6; bnez $v0` where retail used `$at`. I had broadcast that retail materializes `$at` when a comparison constant falls outside the immediate range, so the fix is the FORM OF THE CONSTANT and not the register. Writing the equivalent relational form `selector > 5` made MWCC select `$at`. That is the second function this wave closed by reshaping a comparison constant, after `k_sceneDraw 0019F8F0` fell to a `> 0xff` clamp.

**003C8B50's nd2 was not coloring at all -- it was two WRONG BRANCH DESTINATIONS.** `fclMisc9390Call` and `fclMisc9340Call` were scoped inside the flag-1 block and the flag-4 store was unconditional; correcting the scopes reproduced both retail edges. Worth remembering that a very low nd can still be a control-flow bug, not a register one -- read the branch TARGETS in the diff, not just the opcodes.

**0021C9F0 was a struct-shape error.** We treated `texture+8` as an inline array (`sll v1; addiu v1,8; addu v1,s0,v1`) where retail loads a POINTER from `8(s0)` and indexes it (`sll a0; lw v1,8(s0); addu v1,v1,a0`). Adding the typed texture-header layout and going through its raster-table pointer closed it. That is the struct layer paying for byte-exactness rather than merely for readability.

SIX UNRESOLVED, each driven to an exact instruction-level cause and left honest. Recording them so nobody re-derives:
  frTagTable 003AF660 nd1   offset 0x24 `addu $v1,$v1,$v0` vs retail `addu $v1,$v0,$v1`. Swapping assignment order was byte-neutral; folding the field load into the pointer expression swapped the two preceding loads and regressed to nd3.
  h_pad      00104140 nd4   switch lowering uses `$v0` for 0x27d8 at 0x50/0x54 instead of retail `$v1`, then an unconditional `b` at 0x68 instead of retail's redundant `bne $s1,$v1`.
  k_clump    001A7370 nd5   first loop only: table base `$a0` vs retail `$v1`, state[0] `$v1` vs retail `$a0`. Direct global access regressed to nd18; declaration order byte-neutral.
  h_pad      00103F50 nd6   two live roles globally swapped -- interrupt state `$s0` vs retail `$s1`, copy size `$s1` vs retail `$s0`.
  btlAction  00297480 nd6   only adjacent call-argument load order: we do `lw $a0,0x30($s1)` then `lhu $a1,0x74($s1)`; retail does the `lhu` first.
  mdlManager 003197C0 nd7   inner-loop count/copy-word roles swapped -- constant 8 in `$a3` vs retail `$a2`. Declaration swap byte-neutral; moving the count ahead of the outer index regressed to nd20.

A TOOLING NOTE worth propagating: for functions that have been RENAMED via `#pragma alias`, a literal `FUN_<addr>` lookup in fndiff resolves no marker name. Rerun with the actual C symbol plus an explicit `--addr`.

451/598 across the eight files, no lost MATCHes, over-window identical at 14 with no additions, 598 markers before and after, all pragma brackets balanced, LF/CRLF preserved per file, and scoped decomp_lint introduced ZERO new findings in any of the eight.

+++ src/Battle/btlCamera.c
@@ -414,7 +414,7 @@ void btlCameraSetState(u16 state, BtlAction* action, u32 param_3)
     temp_6 = state & 0xffff;
     if (gBtl->camera.state != temp_6 || entry->unk_08 != 0)
     {
-        var_7 = 2;
+        var_7 = (u64)2;
         while ((u16)var_7 == 0)
         {
             *(u16*)(iGpffffb6fc + 0x104 + (u16)var_7 * 2) =
